# openapi.json
{
  "components": {
    "schemas": {
...
      "AuthenticatedUser": {
        "additionalProperties": false,
        "description": "The signed-in user.",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "displayName"
        ],
        "type": "object"
      },
...
      "CreateOrgRequestInput": {
        "additionalProperties": false,
        "description": "Creates an organization with the calling user as its Owner. No slug — it is derived from the name.",
        "properties": {
          "fiscalYearStartMonth": {
            "description": "The calendar month the org’s fiscal year begins in, 1–12. April, July, and October are all common. Defaults to January where the field is optional.",
            "maximum": 12,
            "minimum": 1,
            "type": "integer"
          },
          "name": {
            "description": "Length is bounded by the server (`orgs.name` is `VARCHAR(255)`) and reported as a `validation_failed` naming `name`.",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "The body of every non-2xx response.",
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "description": "Stable machine-readable failure code. Integrators branch on this, not on the HTTP status: `conflict` and `idempotency_key_conflict` are both 409 and have different recoveries. Codes are never renamed and never change status.",
                "enum": [
                  "validation_failed",
                  "unauthenticated",
                  "permission_denied",
                  "not_found",
                  "conflict",
                  "idempotency_key_conflict",
                  "precondition_failed",
                  "internal_error"
                ],
                "type": "string"
              },
              "details": {
                "additionalProperties": {},
                "propertyNames": {
                  "type": "string"
                },
                "type": "object"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Identity": {
        "additionalProperties": false,
        "description": "The caller, every organization they are a member of, and which one this session is currently scoped to. `activeOrgId` is null for a user who is a member of nothing.",
        "properties": {
          "activeOrgId": {
            "anyOf": [
              {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "memberships": {
            "items": {
              "$ref": "#/components/schemas/OrgMembership"
            },
            "type": "array"
          },
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUser"
          }
        },
        "required": [
          "user",
          "memberships",
          "activeOrgId"
        ],
        "type": "object"
      },
...
      "LoginRequestInput": {
        "additionalProperties": false,
        "description": "Exchanges a password for a session cookie. Reachable without credentials. Every failure answers `unauthenticated` with no detail — a wrong password and an unknown address are indistinguishable by design.",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
...
      "OrgMembership": {
        "additionalProperties": false,
        "description": "One organization together with the role the caller holds in it.",
        "properties": {
          "org": {
            "$ref": "#/components/schemas/OrgSummary"
          },
          "roleCode": {
            "type": "string"
          },
          "roleId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          }
        },
        "required": [
          "org",
          "roleId",
          "roleCode"
        ],
        "type": "object"
      },
...
      "OrgSummary": {
        "additionalProperties": false,
        "description": "One organization the caller can act in.",
        "properties": {
          "fiscalYearStartMonth": {
            "description": "The calendar month the org’s fiscal year begins in, 1–12. April, July, and October are all common. Defaults to January where the field is optional.",
            "maximum": 12,
            "minimum": 1,
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "fiscalYearStartMonth"
        ],
        "type": "object"
      },
...
      "RegisterRequestInput": {
        "additionalProperties": false,
        "description": "Creates a user, their first organization, an Owner membership, and a session — atomically. Reachable without credentials.",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "org": {
            "$ref": "#/components/schemas/CreateOrgRequestInput"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "displayName",
          "org"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Open-source, AI-native, modular double-entry accounting. Every capability in the product is reachable through this API (spec §8) — the web client is a consumer of it, not a privileged path.\n\nMonetary amounts are strings of minor units (`\"1999\"` is 19.99 in a two-decimal currency), never JSON numbers. A JSON number cannot represent a cent exactly and any client that parses one has already lost the value.\n\nEvery write endpoint requires an `Idempotency-Key` header. Reusing a key with the same request replays the original outcome; reusing it with a different request is refused with `idempotency_key_conflict`.",
    "license": {
      "identifier": "AGPL-3.0-only",
      "name": "AGPL-3.0-only"
    },
    "title": "OpenBooks API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "description": "Sets an `HttpOnly` session cookie. Every failure answers `unauthenticated` with no detail and takes the same time, because a faster answer for an unknown address is a user-enumeration oracle regardless of what the body says. Reachable without credentials.",
        "operationId": "login",
        "parameters": [
          {
            "description": "Required on every write (spec §12). Send one unique value per logical request and reuse it verbatim when retrying: the same key with the same request replays the original outcome, and the same key with a different request is refused with `idempotency_key_conflict`.",
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Identity"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Exchange a password for a session",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "description": "Revokes the session named by the cookie and clears it. Succeeds whether or not the cookie names a live session — a stale cookie is the ordinary case for this call.",
        "operationId": "logout",
        "parameters": [
          {
            "description": "Required on every write (spec §12). Send one unique value per logical request and reuse it verbatim when retrying: the same key with the same request replays the original outcome, and the same key with a different request is refused with `idempotency_key_conflict`.",
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Revoke the current session",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "description": "Answers for any live session, including one whose user is a member of no organization — `activeOrgId` is then null.",
        "operationId": "getCurrentIdentity",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Identity"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "The caller, their organizations, and the active one",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/register": {
      "post": {
        "description": "Creates a user, their first organization, an Owner membership, and a session, atomically. Sets an `HttpOnly` session cookie; the token is never in the body. Reachable without credentials.",
        "operationId": "register",
        "parameters": [
          {
            "description": "Required on every write (spec §12). Send one unique value per logical request and reuse it verbatim when retrying: the same key with the same request replays the original outcome, and the same key with a different request is refused with `idempotency_key_conflict`.",
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Identity"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Register a user and their first organization",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
...
}